Added /auth route
