+++ Client/routes/index.js
@@ -33,7 +33,7 @@ router.get('/profile', function(req, res, next)
   }
   else
   {
-    res.redirect('/login');
+    res.redirect('/auth/login?url=/profile');
   }
 });
