+++ Client/views/pages/register.ejs
@@ -26,7 +26,7 @@
               <span class="form-error">Please enter the same password as you did previously.</span>
             </label>
             <p><button class="button expanded">Register</button></p>
-            <p class="text-center"><a href="/login">Already have an account? Log in.</a></p>   
+            <p class="text-center"><a href="/auth/login">Already have an account? Log in.</a></p>   
           </div>
         </form>
