+++ Client/views/partials/head.ejs
@@ -12,5 +12,5 @@
     <![endif]-->
 
     <!-- styles -->
-    <link href="css/app.css" rel="stylesheet">
+    <link href="/css/app.css" rel="stylesheet">
 </head>
