+++ Client/views/partials/scripts.ejs
@@ -1,6 +1,6 @@
 <!-- views/partials/scripts.ejs -->
 
-<script src="components/jquery/dist/jquery.min.js"></script>
-<script src="components/what-input/what-input.min.js"></script>
-<script src="components/foundation-sites/dist/foundation.min.js"></script>
-<script src="js/app.js"></script>
+<script src="/components/jquery/dist/jquery.min.js"></script>
+<script src="/components/what-input/what-input.min.js"></script>
+<script src="/components/foundation-sites/dist/foundation.min.js"></script>
+<script src="/js/app.js"></script>
